+++ packages/playground/ssr-vue/src/pages/Home.vue
@@ -6,18 +6,21 @@
   <button @click="state.count++">count is: {{ state.count }}</button>
   <Foo />
   <p class="virtual">msg from virtual module: {{ foo.msg }}</p>
+  <p class="inter">this will be styled with a font-face</p>
 
-  <ImportType/>
+  <ImportType />
 </template>
 
 <script setup>
 import foo from '@foo'
 import { reactive, defineAsyncComponent } from 'vue'
-import ImportType from '../components/ImportType.vue'
+const ImportType = load('ImportType')
 const Foo = defineAsyncComponent(() =>
   import('../components/Foo').then((mod) => mod.Foo)
 )
-
+function load(file) {
+  return defineAsyncComponent(() => import(`../components/${file}.vue`))
+}
 const state = reactive({ count: 0 })
 </script>