+++ packages/playground/vue/CustomBlock.vue
@@ -3,12 +3,12 @@
   <p class="custom-block">{{ t('hello') }}</p>
 </template>
 
-<script>
+<script lang="ts">
 import { getCurrentInstance } from 'vue'
 
 function useI18n(locale = 'en') {
   const instance = getCurrentInstance()
-  const resources = instance.type.i18n || { en: {} }
+  const resources = (instance.type as any).i18n || { en: {} }
   function t(key) {
     const res = resources[locale] || {}
     return res[key]
@@ -16,6 +16,7 @@ function useI18n(locale = 'en') {
   return { t }
 }
 
+// export default
 export default {
   setup() {
     return { ...useI18n('ja') }