+++ packages/vite/client.d.ts
@@ -25,18 +25,14 @@ interface ImportMeta {
 
   readonly env: ImportMetaEnv
 
-  glob(
-    pattern: string
-  ): Record<
+  glob(pattern: string): Record<
     string,
     () => Promise<{
       [key: string]: any
     }>
   >
 
-  globEager(
-    pattern: string
-  ): Record<
+  globEager(pattern: string): Record<
     string,
     {
       [key: string]: any
@@ -78,6 +74,10 @@ declare module '*.module.stylus' {
   const classes: CSSModuleClasses
   export default classes
 }
+declare module '*.module.pcss' {
+  const classes: CSSModuleClasses
+  export default classes
+}
 
 // CSS
 declare module '*.css' {
@@ -101,6 +101,10 @@ declare module '*.styl' {
   export default css
 }
 declare module '*.stylus' {
+  const css: string
+  export default css
+}
+declare module '*.pcss' {
   const css: string
   export default css
 }
@@ -203,6 +207,13 @@ declare module '*?worker&inline' {
     new (): Worker
   }
   export default workerConstructor
+}
+
+declare module '*?sharedworker' {
+  const sharedWorkerConstructor: {
+    new (): SharedWorker
+  }
+  export default sharedWorkerConstructor
 }
 
 declare module '*?raw' {