+++ packages/vite/src/node/build.ts
@@ -51,9 +51,8 @@ export interface BuildOptions {
    * import)
    *
    * Default: 'modules' - Similar to `@babel/preset-env`'s targets.esmodules,
-   * transpile targeting browsers that natively support es module imports. Also
-   * injects a light-weight dynamic import polyfill.
-   * https://caniuse.com/es6-module
+   * transpile targeting browsers that natively support dynamic es module imports.
+   * https://caniuse.com/es6-module-dynamic-import
    *
    * Another special value is 'esnext' - which only performs minimal transpiling
    * (for minification compat) and assumes native dynamic imports support.
@@ -63,9 +62,9 @@ export interface BuildOptions {
    */
   target?: 'modules' | TransformOptions['target'] | false
   /**
-   * Whether to inject dynamic import polyfill. Defaults to `true`, unless
-   * `target` is `'esnext'`.
+   * whether to inject dynamic import polyfill.
    * Note: does not apply to library mode.
+   * @default false
    */
   polyfillDynamicImport?: boolean
   /**
@@ -200,7 +199,7 @@ export type ResolvedBuildOptions = Required<Omit<BuildOptions, 'base'>>
 export function resolveBuildOptions(raw?: BuildOptions): ResolvedBuildOptions {
   const resolved: ResolvedBuildOptions = {
     target: 'modules',
-    polyfillDynamicImport: raw?.target !== 'esnext' && !raw?.lib,
+    polyfillDynamicImport: false,
     outDir: 'dist',
     assetsDir: 'assets',
     assetsInlineLimit: 4096,
@@ -229,10 +228,15 @@ export function resolveBuildOptions(raw?: BuildOptions): ResolvedBuildOptions {
 
   // handle special build targets
   if (resolved.target === 'modules') {
-    // https://caniuse.com/es6-module
-    // edge18 according to js-table (destructuring is not supported in edge16)
-    // https://github.com/evanw/esbuild/blob/d943e89e50696647d6c89ae623ddfdf564ad3cfc/internal/compat/js_table.go#L84
-    resolved.target = ['es2019', 'edge18', 'firefox60', 'chrome61', 'safari11']
+    // Support browserslist
+    // "defaults and supports es6-module and supports es6-module-dynamic-import",
+    resolved.target = [
+      'es2019',
+      'edge88',
+      'firefox78',
+      'chrome87',
+      'safari13.1'
+    ]
   } else if (resolved.target === 'esnext' && resolved.minify !== 'esbuild') {
     // esnext + terser: limit to es2019 so it can be minified by terser
     resolved.target = 'es2019'
@@ -246,14 +250,18 @@ export function resolveBuildOptions(raw?: BuildOptions): ResolvedBuildOptions {
   return resolved
 }
 
-export function resolveBuildPlugins(
-  config: ResolvedConfig
-): { pre: Plugin[]; post: Plugin[] } {
+export function resolveBuildPlugins(config: ResolvedConfig): {
+  pre: Plugin[]
+  post: Plugin[]
+} {
   const options = config.build
   return {
     pre: [
       buildHtmlPlugin(config),
-      commonjsPlugin(options.commonjsOptions),
+      commonjsPlugin({
+        ignoreDynamicRequires: true,
+        ...options.commonjsOptions
+      }),
       dataURIPlugin(),
       dynamicImportVars({
         warnOnError: true,
@@ -335,9 +343,9 @@ async function doBuild(
   const outDir = resolve(options.outDir)
 
   // inject ssr arg to plugin load/transform hooks
-  const plugins = (ssr
-    ? config.plugins.map((p) => injectSsrFlagToHooks(p))
-    : config.plugins) as Plugin[]
+  const plugins = (
+    ssr ? config.plugins.map((p) => injectSsrFlagToHooks(p)) : config.plugins
+  ) as Plugin[]
 
   // inject ssrExternal if present
   const userExternal = options.rollupOptions?.external
@@ -471,13 +479,8 @@ async function doBuild(
       watcher.on('event', (event) => {
         if (event.code === 'BUNDLE_START') {
           config.logger.info(chalk.cyanBright(`\nbuild started...`))
-
-          // clean previous files
           if (options.write) {
-            emptyDir(outDir)
-            if (fs.existsSync(config.publicDir)) {
-              copyDir(config.publicDir, outDir)
-            }
+            prepareOutDir(outDir, options.emptyOutDir, config)
           }
         } else if (event.code === 'BUNDLE_END') {
           event.result.close()
@@ -504,28 +507,7 @@ async function doBuild(
     }
 
     if (options.write) {
-      // warn if outDir is outside of root
-      if (fs.existsSync(outDir)) {
-        const inferEmpty = options.emptyOutDir === null
-        if (
-          options.emptyOutDir ||
-          (inferEmpty && normalizePath(outDir).startsWith(config.root + '/'))
-        ) {
-          emptyDir(outDir)
-        } else if (inferEmpty) {
-          config.logger.warn(
-            chalk.yellow(
-              `\n${chalk.bold(`(!)`)} outDir ${chalk.white.dim(
-                outDir
-              )} is not inside project root and will not be emptied.\n` +
-                `Use --emptyOutDir to override.\n`
-            )
-          )
-        }
-      }
-      if (fs.existsSync(config.publicDir)) {
-        copyDir(config.publicDir, outDir)
-      }
+      prepareOutDir(outDir, options.emptyOutDir, config)
     }
 
     if (Array.isArray(outputs)) {
@@ -543,6 +525,34 @@ async function doBuild(
   }
 }
 
+function prepareOutDir(
+  outDir: string,
+  emptyOutDir: boolean | null,
+  config: ResolvedConfig
+) {
+  if (fs.existsSync(outDir)) {
+    if (
+      emptyOutDir == null &&
+      !normalizePath(outDir).startsWith(config.root + '/')
+    ) {
+      // warn if outDir is outside of root
+      config.logger.warn(
+        chalk.yellow(
+          `\n${chalk.bold(`(!)`)} outDir ${chalk.white.dim(
+            outDir
+          )} is not inside project root and will not be emptied.\n` +
+            `Use --emptyOutDir to override.\n`
+        )
+      )
+    } else if (emptyOutDir !== false) {
+      emptyDir(outDir, ['.git'])
+    }
+  }
+  if (config.publicDir && fs.existsSync(config.publicDir)) {
+    copyDir(config.publicDir, outDir)
+  }
+}
+
 function getPkgName(root: string) {
   const { name } = JSON.parse(lookupFile(root, ['package.json']) || `{}`)
 
@@ -557,14 +567,14 @@ function createMoveToVendorChunkFn(config: ResolvedConfig): GetManualChunk {
     if (
       id.includes('node_modules') &&
       !isCSSRequest(id) &&
-      !hasDynamicImporter(id, getModuleInfo, cache)
+      staticImportedByEntry(id, getModuleInfo, cache)
     ) {
       return 'vendor'
     }
   }
 }
 
-function hasDynamicImporter(
+function staticImportedByEntry(
   id: string,
   getModuleInfo: GetModuleInfo,
   cache: Map<string, boolean>,
@@ -583,15 +593,21 @@ function hasDynamicImporter(
     cache.set(id, false)
     return false
   }
-  if (mod.dynamicImporters.length) {
+
+  if (mod.isEntry) {
     cache.set(id, true)
     return true
   }
-  const someImporterHas = mod.importers.some((importer) =>
-    hasDynamicImporter(importer, getModuleInfo, cache, importStack.concat(id))
+  const someImporterIs = mod.importers.some((importer) =>
+    staticImportedByEntry(
+      importer,
+      getModuleInfo,
+      cache,
+      importStack.concat(id)
+    )
   )
-  cache.set(id, someImporterHas)
-  return someImporterHas
+  cache.set(id, someImporterIs)
+  return someImporterIs
 }
 
 function resolveBuildOutputs(