+++ packages/vite/src/node/index.ts
@@ -12,8 +12,10 @@ export type {
   ViteDevServer,
   ServerOptions,
   CorsOptions,
+  FileSystemServeOptions,
   CorsOrigin,
-  ServerHook
+  ServerHook,
+  ResolvedServerOptions
 } from './server'
 export type {
   BuildOptions,