+++ packages/vite/src/node/plugins/clientInjections.ts
@@ -25,10 +25,10 @@ export function clientInjectionsPlugin(config: ResolvedConfig): Plugin {
         const overlay = options.overlay !== false
         let port
         if (config.server.middlewareMode) {
-          port = String(
-            (typeof config.server.hmr === 'object' && config.server.hmr.port) ||
-              24678
-          )
+          if (typeof config.server.hmr === 'object') {
+            port = config.server.hmr.clientPort || config.server.hmr.port
+          }
+          port = String(port || 24678)
         } else {
           port = String(options.port || config.server.port!)
         }