+++ packages/vite/src/node/plugins/importAnalysis.ts
@@ -112,6 +112,10 @@ export function importAnalysisPlugin(config: ResolvedConfig): Plugin {
       const rewriteStart = Date.now()
       await init
       let imports: readonly ImportSpecifier[] = []
+      // strip UTF-8 BOM
+      if (source.charCodeAt(0) === 0xfeff) {
+        source = source.slice(1)
+      }
       try {
         imports = parseImports(source)[0]
       } catch (e) {
@@ -284,29 +288,28 @@ export function importAnalysisPlugin(config: ResolvedConfig): Plugin {
           } else if (prop === '.glo' && source[end + 4] === 'b') {
             // transform import.meta.glob()
             // e.g. `import.meta.glob('glob:./dir/*.js')`
-            const {
-              imports,
-              importsString,
-              exp,
-              endIndex,
-              base,
-              pattern
-            } = await transformImportGlob(
-              source,
-              start,
-              importer,
-              index,
-              root,
-              normalizeUrl
-            )
+            const { imports, importsString, exp, endIndex, base, pattern } =
+              await transformImportGlob(
+                source,
+                start,
+                importer,
+                index,
+                root,
+                normalizeUrl
+              )
             str().prepend(importsString)
             str().overwrite(expStart, endIndex, exp)
             imports.forEach((url) => importedUrls.add(url.replace(base, '/')))
-            server._globImporters[importerModule.file!] = {
-              module: importerModule,
+            if (!(importerModule.file! in server._globImporters)) {
+              server._globImporters[importerModule.file!] = {
+                module: importerModule,
+                importGlobs: []
+              }
+            }
+            server._globImporters[importerModule.file!].importGlobs.push({
               base,
               pattern
-            }
+            })
           }
           continue
         }
@@ -540,10 +543,12 @@ function transformCjsImport(
   rawUrl: string,
   importIndex: number
 ): string | undefined {
-  const node = (parseJS(importExp, {
-    ecmaVersion: 2020,
-    sourceType: 'module'
-  }) as any).body[0] as Node
+  const node = (
+    parseJS(importExp, {
+      ecmaVersion: 2020,
+      sourceType: 'module'
+    }) as any
+  ).body[0] as Node
 
   if (node.type === 'ImportDeclaration') {
     if (!node.specifiers.length) {