+++ packages/vite/src/node/plugins/importAnalysisBuild.ts
@@ -121,9 +121,12 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
       let needPreloadHelper = false
 
       for (let index = 0; index < imports.length; index++) {
-        const { s: start, e: end, ss: expStart, d: dynamicIndex } = imports[
-          index
-        ]
+        const {
+          s: start,
+          e: end,
+          ss: expStart,
+          d: dynamicIndex
+        } = imports[index]
 
         const isGlob =
           source.slice(start, end) === 'import.meta' &&
@@ -131,20 +134,16 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
 
         // import.meta.glob
         if (isGlob) {
-          const {
-            importsString,
-            exp,
-            endIndex,
-            isEager
-          } = await transformImportGlob(
-            source,
-            start,
-            importer,
-            index,
-            config.root,
-            undefined,
-            ssr
-          )
+          const { importsString, exp, endIndex, isEager } =
+            await transformImportGlob(
+              source,
+              start,
+              importer,
+              index,
+              config.root,
+              undefined,
+              ssr
+            )
           str().prepend(importsString)
           str().overwrite(expStart, endIndex, exp)
           if (!isEager) {
@@ -221,7 +220,6 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
 
           if (imports.length) {
             const s = new MagicString(code)
-            const analyzed: Set<string> = new Set<string>()
             for (let index = 0; index < imports.length; index++) {
               const { s: start, e: end, d: dynamicIndex } = imports[index]
               // if dynamic import polyfill is used, rewrite the import to
@@ -236,6 +234,7 @@ export function buildImportAnalysisPlugin(config: ResolvedConfig): Plugin {
               if (url[0] === `"` && url[url.length - 1] === `"`) {
                 const ownerFilename = chunk.fileName
                 // literal import - trace direct imports and add to deps
+                const analyzed: Set<string> = new Set<string>()
                 const addDeps = (filename: string) => {
                   if (filename === ownerFilename) return
                   if (analyzed.has(filename)) return