+++ packages/vite/src/node/plugins/worker.ts
@@ -24,8 +24,14 @@ export function webWorkerPlugin(config: ResolvedConfig): Plugin {
     name: 'vite:worker',
 
     load(id) {
-      if (isBuild && parseWorkerRequest(id)?.worker != null) {
-        return ''
+      if (isBuild) {
+        const parsedQuery = parseWorkerRequest(id)
+        if (
+          parsedQuery &&
+          (parsedQuery.worker ?? parsedQuery.sharedworker) != null
+        ) {
+          return ''
+        }
       }
     },
 
@@ -36,14 +42,17 @@ export function webWorkerPlugin(config: ResolvedConfig): Plugin {
           code: `import '${ENV_PUBLIC_PATH}'\n` + _
         }
       }
-      if (query == null || (query && query.worker == null)) {
+      if (
+        query == null ||
+        (query && (query.worker ?? query.sharedworker) == null)
+      ) {
         return
       }
 
       let url: string
       if (isBuild) {
         if (query.inline != null) {
-          // bundle the file as entry to support imports and inline as base64
+          // bundle the file as entry to support imports and inline as blob
           // data url
           const rollup = require('rollup') as typeof Rollup
           const bundle = await rollup.rollup({
@@ -55,11 +64,18 @@ export function webWorkerPlugin(config: ResolvedConfig): Plugin {
               format: 'es',
               sourcemap: config.build.sourcemap
             })
-            url = `data:application/javascript;base64,${Buffer.from(
-              output[0].code
-            ).toString('base64')}`
+            
+            return `const blob = new Blob([atob(\"${Buffer.from(output[0].code).toString('base64')}\")], { type: 'text/javascript;charset=utf-8' });
+            export default function WorkerWrapper() {
+              const objURL = (window.URL || window.webkitURL).createObjectURL(blob);
+              try {
+                return new Worker(objURL);
+              } finally {
+                (window.URL || window.webkitURL).revokeObjectURL(objURL);
+              }
+            }`
           } finally {
-            bundle.close()
+            await bundle.close()
           }
         } else {
           // emit as separate chunk
@@ -73,8 +89,14 @@ export function webWorkerPlugin(config: ResolvedConfig): Plugin {
         url = injectQuery(url, WorkerFileId)
       }
 
+      const workerConstructor =
+        query.sharedworker != null ? 'SharedWorker' : 'Worker'
+      const workerOptions = { type: 'module' }
+
       return `export default function WorkerWrapper() {
-        return new Worker(${JSON.stringify(url)}, { type: 'module' })
+        return new ${workerConstructor}(${JSON.stringify(
+        url
+      )}, ${JSON.stringify(workerOptions, null, 2)})
       }`
     }
   }