+++ packages/vite/src/node/server/middlewares/transform.ts
@@ -58,7 +58,8 @@ export function transformMiddleware(
     }
   }
 
-  return async (req, res, next) => {
+  // Keep the named function. The name is visible in debug logs via `DEBUG=connect:dispatcher ...`
+  return async function viteTransformMiddleware(req, res, next) {
     if (req.method !== 'GET' || knownIgnoreList.has(req.url!)) {
       return next()
     }