+++ scripts/jestPerTestSetup.ts
@@ -1,15 +1,19 @@
 import fs from 'fs-extra'
 import * as http from 'http'
 import { resolve, dirname } from 'path'
-import slash from 'slash'
 import sirv from 'sirv'
 import { createServer, build, ViteDevServer, UserConfig } from 'vite'
 import { Page } from 'playwright-chromium'
 
 const isBuildTest = !!process.env.VITE_TEST_BUILD
 
+export function slash(p: string): string {
+  return p.replace(/\\/g, '/')
+}
+
 // injected by the test env
 declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace NodeJS {
     interface Global {
       page?: Page
@@ -66,18 +70,21 @@ beforeAll(async () => {
 
       const options: UserConfig = {
         root: tempDir,
-        logLevel: 'error',
+        logLevel: 'silent',
         server: {
           watch: {
             // During tests we edit the files too fast and sometimes chokidar
             // misses change events, so enforce polling for consistency
             usePolling: true,
             interval: 100
+          },
+          host: true,
+          fsServe: {
+            strict: !isBuildTest
           }
         },
         build: {
-          // skip transpilation and dynamic import polyfills during tests to
-          // make it faster
+          // skip transpilation during tests to make it faster
           target: 'esnext'
         }
       }
@@ -87,7 +94,8 @@ beforeAll(async () => {
         server = await (await createServer(options)).listen()
         // use resolved port/base from server
         const base = server.config.base === '/' ? '' : server.config.base
-        const url = (global.viteTestUrl = `http://localhost:${server.config.server.port}${base}`)
+        const url =
+          (global.viteTestUrl = `http://localhost:${server.config.server.port}${base}`)
         await page.goto(url)
       } else {
         process.env.VITE_INLINE = 'inline-build'
@@ -100,14 +108,19 @@ beforeAll(async () => {
     // jest doesn't exit if our setup has error here
     // https://github.com/facebook/jest/issues/2713
     err = e
+
+    // Closing the page since an error in the setup, for example a runtime error
+    // when building the playground should skip further tests.
+    // If the page remains open, a command like `await page.click(...)` produces
+    // a timeout with an exception that hides the real error in the console.
+    await page.close()
   }
 }, 30000)
 
 afterAll(async () => {
-  global.page && global.page.off('console', onConsole)
-  if (server) {
-    await server.close()
-  }
+  global.page?.off('console', onConsole)
+  await global.page?.close()
+  await server?.close()
   if (err) {
     throw err
   }